IMPORTANT: few modifications into can.h (use of strict sized types). >>> Need to be tested <<<
Big commit containing a very first try of the FSUIPC <-> CAN management.

Offsets are represented as Data and implemented by OffsetAccessor.
FsAdapter deals with changes from FSUIPC and send CAN message if needed.
It deals also with CAN messages input and apply changes to FSUIPC.

Code is going to be documented as short as possible (tomorrow).

--Julien.